# src/main/java/vvu/centrauthz/domains/resources/controllers/ResourceController.java
package vvu.centrauthz.domains.resources.controllers;

import io.micronaut.http.HttpResponse;
import io.micronaut.http.annotation.*;
import io.micronaut.core.annotation.NonNull;
import reactor.core.publisher.Mono;
import vvu.centrauthz.domains.resources.models.Resource;
import vvu.centrauthz.domains.resources.models.ResourceForPatch;
import vvu.centrauthz.exceptions.AppError;

import java.util.UUID;

/**
 * Reactive RESTful API controller for resource management.
 * All endpoints currently return NOT_IMPLEMENTED errors.
 * Validation is handled by Micronaut validation framework.
 */
@Controller("/v0/applications/{applicationKey}/resources")
public class ResourceController {

    /**
     * Get a resource by ID.
     * 
     * @param applicationKey Application key (minimum 3 characters)
     * @param id Unique identifier for the resource
     * @return Mono containing the resource
     */
    @Get("/{id}")
    public Mono<HttpResponse<Resource>> getResource(
            @PathVariable @NonNull String applicationKey,
            @PathVariable @NonNull UUID id) {
        
        return Mono.error(new AppError("NOT_IMPLEMENTED", "Get resource operation is not yet implemented"));
    }

    /**
     * Update a resource.
     * 
     * @param applicationKey Application key (minimum 3 characters)
     * @param id Unique identifier for the resource
     * @param resource Resource data to update
     * @return Mono containing the updated resource
     */
    @Put("/{id}")
    public Mono<HttpResponse<Resource>> updateResource(
            @PathVariable @NonNull String applicationKey,
            @PathVariable @NonNull UUID id,
            @Body @NonNull Resource resource) {
        
        return Mono.error(new AppError("NOT_IMPLEMENTED", "Update resource operation is not yet implemented"));
    }

    /**
     * Partially update a resource.
     * 
     * @param applicationKey Application key (minimum 3 characters)
     * @param id Unique identifier for the resource
     * @param resourcePatch Resource patch data
     * @return Mono containing the updated resource
     */
    @Patch("/{id}")
    public Mono<HttpResponse<Resource>> patchResource(
            @PathVariable @NonNull String applicationKey,
            @PathVariable @NonNull UUID id,
            @Body @NonNull ResourceForPatch resourcePatch) {
        
        return Mono.error(new AppError("NOT_IMPLEMENTED", "Patch resource operation is not yet implemented"));
    }

    /**
     * Delete a resource.
     * 
     * @param applicationKey Application key (minimum 3 characters)
     * @param id Unique identifier for the resource
     * @return Mono indicating completion
     */
    @Delete("/{id}")
    public Mono<HttpResponse<Void>> deleteResource(
            @PathVariable @NonNull String applicationKey,
            @PathVariable @NonNull UUID id) {
        
        return Mono.error(new AppError("NOT_IMPLEMENTED", "Delete resource operation is not yet implemented"));
    }

    /**
     * Create a new resource.
     * 
     * @param applicationKey Application key (minimum 3 characters)
     * @param resource Resource data to create
     * @return Mono containing the created resource
     */
    @Post
    public Mono<HttpResponse<Resource>> createResource(
            @PathVariable @NonNull String applicationKey,
            @Body @NonNull Resource resource) {
        
        return Mono.error(new AppError("NOT_IMPLEMENTED", "Create resource operation is not yet implemented"));
    }
}